server/requirements.txt
Flask
flask-cors
pytest
requests
gunicorn
Werkzeug
python-dotenv
firebase-admin
google-generativeai
google-generativeai>=0.3.2
groq